Moved interpolation script to utils/imputation folder
